Few Changes and Bugfixes
Smartmirror-Config:
Enable or disable features via
config[Calendar,Compliments,Weather,Time,Openhab (not working at the
moment),News]

Requires now an input in certain fields
Only Number items are valid in the interval section on the config page
(Otherwise the config page crashes, when entering a string into the
number sections, or when you leave it empty.
(By default i did not enter any APIKEY in the weather-section, so it
will prompt for an input)
Added description to the items on the config page.
Those changes are implemented in modern Browsers (HTML5)
Tested with Mozilla Firefox, Google Chrome and Microsoft Edge

Smartmirror:
Fixed Weathericons and Datelabel
Prepared support for Openhab (still not working, but i added the
necessary Javascript file, which is a draft at the moment)